start by separating page and page admin tests
